requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
streamlit>=1.28.0
huggingface-hub>=0.16.0
transformers>=4.30.0
timm>=0.9.0
scikit-learn>=1.3.0
moviepy>=1.0.3